test: extract shared registerFakeLauncherApp fixture
AppPickerUtilsTest.kt and GetInstalledAppsUseCaseTest.kt each carried a byte-for-byte
copy of the same ~20-line Robolectric ResolveInfo/ActivityInfo/ApplicationInfo @Before
block; a future PackageManager shadow API change would have had to be applied in both
places or they'd silently drift apart. Also strengthens GetInstalledAppsUseCaseTest's
one test to assert real delegation (structural equality against
getInstalledAppsForPicker()) instead of only shouldNotBeEmpty(), which would have
passed even for an unrelated non-empty result.
