+++ lib/src/test/java/de/lemke/commonutils/AppPickerUtilsTest.kt
@@ -16,10 +16,6 @@
 package de.lemke.commonutils
 
 import android.content.Context
-import android.content.Intent
-import android.content.pm.ActivityInfo
-import android.content.pm.ApplicationInfo
-import android.content.pm.ResolveInfo
 import androidx.picker.di.AppPickerContext
 import androidx.picker.model.AppInfo
 import androidx.picker.model.AppInfoDataImpl
@@ -34,7 +30,6 @@ import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
-import org.robolectric.Shadows.shadowOf
 import org.robolectric.annotation.Config
 
 @RunWith(RobolectricTestRunner::class)
@@ -42,30 +37,8 @@ import org.robolectric.annotation.Config
 class AppPickerUtilsTest {
     private val context: Context get() = ApplicationProvider.getApplicationContext()
 
-    @Suppress("DEPRECATION")
     @Before
-    fun registerFakeLauncherApp() {
-        val launcherIntent = Intent(Intent.ACTION_MAIN).addCategory(Intent.CATEGORY_LAUNCHER)
-        shadowOf(context.packageManager).setResolveInfosForIntent(
-            launcherIntent,
-            listOf(
-                ResolveInfo().apply {
-                    nonLocalizedLabel = "FakeApp"
-                    activityInfo =
-                        ActivityInfo().apply {
-                            packageName = "de.lemke.commonutils.fakeapp"
-                            name = "de.lemke.commonutils.fakeapp.MainActivity"
-                            applicationInfo =
-                                ApplicationInfo().apply {
-                                    packageName = "de.lemke.commonutils.fakeapp"
-                                    nonLocalizedLabel = "FakeApp"
-                                    flags = ApplicationInfo.FLAG_INSTALLED
-                                }
-                        }
-                },
-            ),
-        )
-    }
+    fun registerFakeLauncherApp() = registerFakeLauncherApp(context)
 
     @Test
     fun `convert sets searchable to label and packageName`() {
